added ADMIN role(to add articles); rest api init; in start generate ADMIN root; deleted csrf

+++ manager/src/main/java/server/entities/Article.java
@@ -8,13 +8,14 @@ public class Article
 {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Integer id;
+  private Long id;
 
   @Column(unique = true)
   private String name;
 
   public Article()
   {
+
   }
 
   public Article(String name)
@@ -22,12 +23,12 @@ public class Article
     this.name = name;
   }
 
-  public Integer getId()
+  public Long getId()
   {
     return id;
   }
 
-  public void setId(Integer id)
+  public void setId(Long id)
   {
     this.id = id;
   }
